chore: remove dead code and dedupe across server and frontend
Five independent cleanups flagged by prior audits, each no-behavior-change:

- Extract the duplicated, security-critical sanitizeExternalUrlHref URL
  scheme allowlist from the Note and Event ActivityStreams object parsers
  into a shared activitypub/helper/url-sanitizer module, removing the
  manual "MUST stay in sync" obligation on a security barrier.
- Remove the dead getSharedEventIds method from ActivityPubInterface
  (superseded by getSharedEventStatusMap, no production caller) and drop
  its now-orphaned stubs from the calendar and AP test mocks.
- Drop duplicate event_editor locale imports and resource keys in the
  client locale loader that produced build-time duplicate-key warnings.
- Document the FK invariant the isSameHostAsActor outbox delivery guard
  defends, so its rarity and purpose are explicit.
- Remove the unused calendar prop from the site EventDetailBody component,
  its two shell callers, and the test scaffolding/assertions that
  forwarded it.
