+++ src/server/calendar/test/EventService/replace_event_categories.test.ts
@@ -35,7 +35,6 @@ function buildMockApInterface(
     return calendarIdsForEvent[eventId] ?? [];
   });
   return {
-    getSharedEventIds: sinon.stub().resolves(sharedEventIds),
     getSharedEventStatusMap: sinon.stub().resolves(statusMap),
     getCalendarIdsForSharedEvent,
   } as any;
