+++ src/server/calendar/test/integration/bulk_category_assignment.test.ts
@@ -33,7 +33,6 @@ describe('CalendarInterface.bulkAssignCategories', () => {
     eventBus = new EventEmitter();
     calendarInterface = new CalendarInterface(eventBus);
     calendarInterface.setActivityPubInterface({
-      getSharedEventIds: async () => [],
       getSharedEventStatusMap: async () => new Map<string, 'auto' | 'manual'>(),
       getCalendarIdsForSharedEvent: async () => [],
       getEventSourceActorUris: async () => new Map<string, string>(),
