+++ src/server/calendar/test/integration/event_accessibility_info.test.ts
@@ -31,7 +31,6 @@ describe('Event content accessibilityInfo — API round-trip (integration)', () 
     calendarInterface = new CalendarInterface(eventBus);
 
     calendarInterface.setActivityPubInterface({
-      getSharedEventIds: async () => [],
       getSharedEventStatusMap: async () => new Map(),
       findCalendarActorByCalendarId: async () => null,
     } as any);
