+++ src/server/calendar/test/integration/event_external_url.test.ts
@@ -37,7 +37,6 @@ describe('Event externalUrl + urlPrompt — API round-trip (integration)', () =>
     // Minimal AP interface stub: integration tests here do not wire the AP
     // domain, so we return shapes that keep local-only code paths correct.
     calendarInterface.setActivityPubInterface({
-      getSharedEventIds: async () => [],
       getSharedEventStatusMap: async () => new Map(),
       findCalendarActorByCalendarId: async () => null,
     } as any);
